feature: update todo-next preview.js for storybook
